+++ BE/src/stock/index/stock-index.module.ts
@@ -3,10 +3,11 @@ import { StockIndexController } from './stock-index.controller';
 import { StockIndexService } from './stock-index.service';
 import { KoreaInvestmentModule } from '../../koreaInvestment/korea-investment.module';
 import { SocketModule } from '../../websocket/socket.module';
+import { StockIndexSocketService } from './stock-index-socket.service';
 
 @Module({
   imports: [KoreaInvestmentModule, SocketModule],
   controllers: [StockIndexController],
-  providers: [StockIndexService],
+  providers: [StockIndexService, StockIndexSocketService],
 })
 export class StockIndexModule {}
