+++ BE/src/stock/order/stock-order.module.ts
@@ -1,4 +1,4 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { StockOrderController } from './stock-order.controller';
 import { StockOrderService } from './stock-order.service';
@@ -6,15 +6,11 @@ import { Order } from './stock-order.entity';
 import { StockOrderRepository } from './stock-order.repository';
 import { SocketModule } from '../../websocket/socket.module';
 import { AssetModule } from '../../asset/asset.module';
+import { StockOrderSocketService } from './stock-order-socket.service';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Order]),
-    forwardRef(() => SocketModule),
-    AssetModule,
-  ],
+  imports: [TypeOrmModule.forFeature([Order]), SocketModule, AssetModule],
   controllers: [StockOrderController],
-  providers: [StockOrderService, StockOrderRepository],
-  exports: [StockOrderService],
+  providers: [StockOrderService, StockOrderRepository, StockOrderSocketService],
 })
 export class StockOrderModule {}
