+++ BE/src/websocket/socket.module.ts
@@ -1,21 +1,10 @@
-import { forwardRef, Module } from '@nestjs/common';
-import { StockIndexSocketService } from './stock/index/stock-index-socket.service';
+import { Module } from '@nestjs/common';
 import { SocketGateway } from './socket.gateway';
 import { SocketTokenService } from './socket-token.service';
-import { StockPriceSocketService } from './stock/price/stock-price-socket.service';
 import { BaseSocketService } from './base-socket.service';
-import { StockOrderModule } from '../stock/order/stock-order.module';
 
 @Module({
-  imports: [forwardRef(() => StockOrderModule)],
-  controllers: [],
-  providers: [
-    SocketTokenService,
-    StockIndexSocketService,
-    SocketGateway,
-    StockPriceSocketService,
-    BaseSocketService,
-  ],
-  exports: [SocketGateway, StockPriceSocketService],
+  providers: [SocketTokenService, SocketGateway, BaseSocketService],
+  exports: [SocketGateway, BaseSocketService],
 })
 export class SocketModule {}
