--- BE/src/websocket/stock/price/stock-price-socket.service.ts
@@ -1,51 +1,0 @@
-import {
-  forwardRef,
-  Inject,
-  Injectable,
-  InternalServerErrorException,
-} from '@nestjs/common';
-import { BaseSocketService } from '../../base-socket.service';
-import { SocketGateway } from '../../socket.gateway';
-import { StockOrderService } from '../../../stock/order/stock-order.service';
-import { Order } from '../../../stock/order/stock-order.entity';
-
-@Injectable()
-export class StockPriceSocketService {
-  private TR_ID = 'H0STCNT0';
-
-  constructor(
-    private readonly socketGateway: SocketGateway,
-    private readonly baseSocketService: BaseSocketService,
-    @Inject(forwardRef(() => StockOrderService))
-    private readonly stockOrderService: StockOrderService,
-  ) {
-    baseSocketService.registerSocketOpenHandler(async () => {
-      const orders: Order[] = await stockOrderService.findAllPendingOrderCode();
-      orders.forEach((order) => {
-        baseSocketService.registerCode(this.TR_ID, order.stock_code);
-      });
-    });
-
-    baseSocketService.registerSocketDataHandler(
-      this.TR_ID,
-      (data: string[]) => {
-        stockOrderService
-          .checkExecutableOrder(
-            data[0], // 주식 코드
-            data[2], // 주식 체결가
-          )
-          .catch(() => {
-            throw new InternalServerErrorException();
-          });
-      },
-    );
-  }
-
-  subscribeByCode(trKey: string) {
-    this.baseSocketService.registerCode(this.TR_ID, trKey);
-  }
-
-  unsubscribeByCode(trKey: string) {
-    this.baseSocketService.unregisterCode(this.TR_ID, trKey);
-  }
-}
